UPSTREAM: libpayload/libpci: Introduce device class attribute in pci_dev
The device class is read at different places and it is read from the
hardware directly. Therefore, and in preparation to CB:46416, introduce
the device class attribute in the pci_dev struct. With this, there is
only one interaction with the hardware and it's also more user friendly.

BUG=none
BRANCH=none
TEST=none

Change-Id: I89c13b664cc4b190bae07407744db4ba09d889ef
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 4c9622a7d033e92a214a8534012ff40e2f4bb6e4
Original-Change-Id: I5d56be96f3f0da471246f031ea619e3df8e54cfb
Original-Signed-off-by: Felix Singer <felix.singer@secunet.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/46347
Original-Reviewed-by: Nico Huber <nico.h@gmx.de>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2484475
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>