UPSTREAM: mb/google/volteer: New variant for Volteer reworked with Dauntless
For development of the firmware to run on the Dauntless TPM, a number of
Volteer2 devices are being reworked to replace the H1 chip with probe
wires to connect to an external Dauntless development board.

Some modification to the AP firmware is required, not least because the
Dauntless chip is connected via I2C bus, instead of SPI.  Most of the
Dauntless developers will not otherwise have a Chrome OS chroot.
Because of the above, I think it makes sense to have a new variant, for
the reworked devices, which I intend to create with this CL.

BUG=b:169526865
TEST=abuild -t GOOGLE_VOLTEER2_TI50 -c max -x

Change-Id: I469e3af6a65aa2ca698204e93915c5fd98573b74
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: fbb568347de2037e64e894e2ec68c61ff3de73c8
Original-Change-Id: Ibdcd6c2ce3941c229518f21f0e479890b5d76dd1
Original-Signed-off-by: Jes Bodi Klinke <jbk@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/46437
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2514819
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Jes Klinke <jbk@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>