UPSTREAM: sec/intel/txt: Allow skipping ACM NOP function
This is merely used to test whether the BIOS ACM calling code is working
properly. There's no need to do this on production platforms. Testing on
Haswell showed that running this NOP function breaks S3 resume with TXT.
Add a Kconfig bool to control whether the NOP function is to be invoked.

BUG=none
BRANCH=none
TEST=none

Change-Id: If1aa3bfc50a5902ac0323d5d075408d5236507c0
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8a285fd8a5dabbf6adaf96e4101917ede7c98316
Original-Change-Id: Ibf461c18a96f1add7867e1320726fadec65b7184
Original-Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/46496
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2498313
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>