UPSTREAM: sec/intel/txt: Always run SCHECK on regular boots
When Boot Guard is disabled or not available, the IBB might not even
exist. This is the case on traditional (non-ULT) Haswell, for example.

Leave the S3 resume check as-is for now. Skylake and newer may need to
run SCHECK on resume as well, but I lack the hardware to test this on.

BUG=none
BRANCH=none
TEST=none

Change-Id: I6cd0ce028632f7720f7ef5ff43380eee18263a48
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: e70a3f8822d6c1e0b0f1dc86464acfb24c80b450
Original-Change-Id: I70231f60d4d4c5bc8ee0fcbb0651896256fdd391
Original-Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/46497
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2498314
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>