UPSTREAM: soc/intel/skylake: Use correct NHLT_PDM_DEV definition
According to the NHLT specification[1], PDM_DEV is defined as "1" on
Kabylake based platforms. coreboot currently sets it to "0" on
all platforms. Add an entry to the enum and use it to define
NHLT_PDM_DEV for Kabylake.

"Device Type" will resume from "2" on all platforms, but entries are
currently reserved.

Tested on an Acer Aspire VN7-572G (Skylake-U), which has a 1ch array
DMIC, on Windows 10.

1. https://01.org/sites/default/files/595976_intel_sst_nhlt.pdf

BUG=none
BRANCH=none
TEST=none

Change-Id: Iedbdd373de5c90361ad69fb10771b6f5e735cb3a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 9ee1b82db42623fb700b62d21b5f6d95a0e2e02c
Original-Change-Id: Ifbc67228c9e7af7db5154d597ca8d67860cfd2ed
Original-Signed-off-by: Benjamin Doron <benjamin.doron00@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/45010
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2557632
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>