UPSTREAM: x86: Put bootblock startup code into .text._start section
The initial bootblock assembly code on x86 is just put into the .text
section, which just happens to come before all the individual .text.*
function sections in the program.ld script. So it tends to be at the
start of the image, but if you inserted another linker script section
with contents before .text, it would cause a problem. (I'm not sure if
it's an architectural requirement for _start16bit to come at the start
of the image, but at least its 4K alignment requirement would waste a
lot of space if it didn't.)

This patch moves the section to .text._start which is the name other
architectures use for the code they want in the very front of the image
and which is listed first in program.ld.

BUG=none
BRANCH=none
TEST=none

Change-Id: I94384afb834fcd6e5f715629a0a36457caf06030
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 0ba16637d8f12fe9ba8388222cfa71fc5206c0f3
Original-Signed-off-by: Julius Werner <jwerner@chromium.org>
Original-Change-Id: Ia84e6e33ec29584d356e226e8fdcb8c9334d49af
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/46834
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2577255
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>