UPSTREAM: soc/intel/broadwell: Drop enable check from LPD0/LPD3
When SerialIO devices are disabled, their _STA method evaluates to 0,
which means the device is not present. It is expected that OSPM would
not attempt to change the power state of a device that is not present.
Lynxpoint does not have these checks, thus remove them from Broadwell.
Also remove the now-unused Arg1 parameter to avoid warnings from IASL.

BUG=none
BRANCH=none
TEST=none

Change-Id: I1ad5ef4854c3bbd3a772b9b1463d2f180f1ee312
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ae4eba9be55c8dd660a63ea45eecb2604cfae7ce
Original-Change-Id: Ic5e999ac1171ce49db66bec45c58d8aa5711ec53
Original-Signed-off-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/46966
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Michael Niewöhner <foss@mniewoehner.de>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2647376
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>