UPSTREAM: mb/google/zork: Power off fingerprint sensor on shutdown
When the system shuts down, turn the fingerprint sensor off. This sets
the GPIOs correctly for the next boot.  The fingerprint sensor was
previously left on, and was just powering down when the rails went low.

On suspend, the fingerprint sensor stays awake and puts itself in a low
powerstate mode based on the SLP_Sx_L pin states.

BUG=b:171837716
TEST=Fingerprint sensor still works after S3, GPIO state on the boot
following a shutdown is low.
BRANCH=Zork

Signed-off-by: Martin Roth <gaumless@gmail.com>
Original-Commit-Id: d5c3d9ca82c89830f33ec62dcb15c50ca9ae2523
Original-Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Change-Id: I3837b58372d8f4a504535e76bd21c667d68f8995
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/47311
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Change-Id: I7d483b2b5b6cb9914fc4d0eddf74e50af218e4fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2543974
Reviewed-by: Martin Roth <martinroth@google.com>
Tested-by: Martin Roth <martinroth@google.com>
Commit-Queue: Martin Roth <martinroth@google.com>