UPSTREAM: cpu/x86/smm: Check that the stub size is < save state size
If the stub size would be larger than the save state size, the stagger
points would overlap with the stub.

The check is placed in the stub placement code. The stub placement
code is called twice. Once for the initial SMM relocatation and for
the permanent handler in TSEG. So the check is done twice, which is
not really needed.

BUG=none
BRANCH=none
TEST=none

Change-Id: Id5f43f164658a1accc6c637554f451921ff0e68d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: fd8619e665416fe1d7b7b4c1e69a7ceb1ea58ef8
Original-Change-Id: I253e1a7112cd8f7496cb1a826311f4dd5ccfc73a
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/c/coreboot/+/47069
Original-Reviewed-by: David Hendricks <david.hendricks@gmail.com>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/coreboot/+/2540334
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>
Commit-Queue: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>